# src/gen/java/io/swagger/model/Film.java
/*
 * VideoRentalStore
 * A sample REST API for a VideoRentalStore
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import io.swagger.model.Customer;
import io.swagger.model.FilmKind;
import java.util.Date;
import javax.validation.constraints.*;

/**
 * Film
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2017-05-16T16:06:48.217+02:00")
public class Film   {
  @JsonProperty("filmId")
  private String filmId = null;

  @JsonProperty("title")
  private String title = null;

  @JsonProperty("kind")
  private FilmKind kind = null;

  @JsonProperty("rentedBy")
  private Customer rentedBy = null;

  @JsonProperty("rentedOn")
  private Date rentedOn = null;

  @JsonProperty("deleted")
  private Boolean deleted = null;

  public Film filmId(String filmId) {
    this.filmId = filmId;
    return this;
  }

   /**
   * Get filmId
   * @return filmId
  **/
  @JsonProperty("filmId")
  @ApiModelProperty(value = "")
  public String getFilmId() {
    return filmId;
  }

  public void setFilmId(String filmId) {
    this.filmId = filmId;
  }

  public Film title(String title) {
    this.title = title;
    return this;
  }

   /**
   * Get title
   * @return title
  **/
  @JsonProperty("title")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public String getTitle() {
    return title;
  }

  public void setTitle(String title) {
    this.title = title;
  }

  public Film kind(FilmKind kind) {
    this.kind = kind;
    return this;
  }

   /**
   * Get kind
   * @return kind
  **/
  @JsonProperty("kind")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public FilmKind getKind() {
    return kind;
  }

  public void setKind(FilmKind kind) {
    this.kind = kind;
  }

  public Film rentedBy(Customer rentedBy) {
    this.rentedBy = rentedBy;
    return this;
  }

   /**
   * Get rentedBy
   * @return rentedBy
  **/
  @JsonProperty("rentedBy")
  @ApiModelProperty(value = "")
  public Customer getRentedBy() {
    return rentedBy;
  }

  public void setRentedBy(Customer rentedBy) {
    this.rentedBy = rentedBy;
  }

  public Film rentedOn(Date rentedOn) {
    this.rentedOn = rentedOn;
    return this;
  }

   /**
   * Get rentedOn
   * @return rentedOn
  **/
  @JsonProperty("rentedOn")
  @ApiModelProperty(value = "")
  public Date getRentedOn() {
    return rentedOn;
  }

  public void setRentedOn(Date rentedOn) {
    this.rentedOn = rentedOn;
  }

  public Film deleted(Boolean deleted) {
    this.deleted = deleted;
    return this;
  }

   /**
   * Get deleted
   * @return deleted
  **/
  @JsonProperty("deleted")
  @ApiModelProperty(value = "")
  public Boolean getDeleted() {
    return deleted;
  }

  public void setDeleted(Boolean deleted) {
    this.deleted = deleted;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Film film = (Film) o;
    return Objects.equals(this.filmId, film.filmId) &&
        Objects.equals(this.title, film.title) &&
        Objects.equals(this.kind, film.kind) &&
        Objects.equals(this.rentedBy, film.rentedBy) &&
        Objects.equals(this.rentedOn, film.rentedOn) &&
        Objects.equals(this.deleted, film.deleted);
  }

  @Override
  public int hashCode() {
    return Objects.hash(filmId, title, kind, rentedBy, rentedOn, deleted);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Film {\n");
    
    sb.append("    filmId: ").append(toIndentedString(filmId)).append("\n");
    sb.append("    title: ").append(toIndentedString(title)).append("\n");
    sb.append("    kind: ").append(toIndentedString(kind)).append("\n");
    sb.append("    rentedBy: ").append(toIndentedString(rentedBy)).append("\n");
    sb.append("    rentedOn: ").append(toIndentedString(rentedOn)).append("\n");
    sb.append("    deleted: ").append(toIndentedString(deleted)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}