# src/gen/java/io/swagger/model/RentDesc.java
/*
 * VideoRentalStore
 * A sample REST API for a VideoRentalStore
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import io.swagger.model.FilmRent;
import java.util.ArrayList;
import java.util.List;
import javax.validation.constraints.*;

/**
 * RentDesc
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2017-05-16T16:06:48.217+02:00")
public class RentDesc   {
  @JsonProperty("customerId")
  private String customerId = null;

  @JsonProperty("films")
  private List<FilmRent> films = new ArrayList<FilmRent>();

  public RentDesc customerId(String customerId) {
    this.customerId = customerId;
    return this;
  }

   /**
   * Get customerId
   * @return customerId
  **/
  @JsonProperty("customerId")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public String getCustomerId() {
    return customerId;
  }

  public void setCustomerId(String customerId) {
    this.customerId = customerId;
  }

  public RentDesc films(List<FilmRent> films) {
    this.films = films;
    return this;
  }

  public RentDesc addFilmsItem(FilmRent filmsItem) {
    this.films.add(filmsItem);
    return this;
  }

   /**
   * Get films
   * @return films
  **/
  @JsonProperty("films")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public List<FilmRent> getFilms() {
    return films;
  }

  public void setFilms(List<FilmRent> films) {
    this.films = films;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    RentDesc rentDesc = (RentDesc) o;
    return Objects.equals(this.customerId, rentDesc.customerId) &&
        Objects.equals(this.films, rentDesc.films);
  }

  @Override
  public int hashCode() {
    return Objects.hash(customerId, films);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class RentDesc {\n");
    
    sb.append("    customerId: ").append(toIndentedString(customerId)).append("\n");
    sb.append("    films: ").append(toIndentedString(films)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}