# src/gen/java/io/swagger/model/RentResponse.java
/*
 * VideoRentalStore
 * A sample REST API for a VideoRentalStore
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import javax.validation.constraints.*;

/**
 * RentResponse
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2017-05-16T16:06:48.217+02:00")
public class RentResponse   {
  @JsonProperty("filmId")
  private String filmId = null;

  @JsonProperty("pay")
  private Integer pay = null;

  @JsonProperty("bonusPoints")
  private Integer bonusPoints = null;

  @JsonProperty("message")
  private String message = null;

  public RentResponse filmId(String filmId) {
    this.filmId = filmId;
    return this;
  }

   /**
   * Get filmId
   * @return filmId
  **/
  @JsonProperty("filmId")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public String getFilmId() {
    return filmId;
  }

  public void setFilmId(String filmId) {
    this.filmId = filmId;
  }

  public RentResponse pay(Integer pay) {
    this.pay = pay;
    return this;
  }

   /**
   * Get pay
   * @return pay
  **/
  @JsonProperty("pay")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public Integer getPay() {
    return pay;
  }

  public void setPay(Integer pay) {
    this.pay = pay;
  }

  public RentResponse bonusPoints(Integer bonusPoints) {
    this.bonusPoints = bonusPoints;
    return this;
  }

   /**
   * Get bonusPoints
   * @return bonusPoints
  **/
  @JsonProperty("bonusPoints")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public Integer getBonusPoints() {
    return bonusPoints;
  }

  public void setBonusPoints(Integer bonusPoints) {
    this.bonusPoints = bonusPoints;
  }

  public RentResponse message(String message) {
    this.message = message;
    return this;
  }

   /**
   * Get message
   * @return message
  **/
  @JsonProperty("message")
  @ApiModelProperty(value = "")
  public String getMessage() {
    return message;
  }

  public void setMessage(String message) {
    this.message = message;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    RentResponse rentResponse = (RentResponse) o;
    return Objects.equals(this.filmId, rentResponse.filmId) &&
        Objects.equals(this.pay, rentResponse.pay) &&
        Objects.equals(this.bonusPoints, rentResponse.bonusPoints) &&
        Objects.equals(this.message, rentResponse.message);
  }

  @Override
  public int hashCode() {
    return Objects.hash(filmId, pay, bonusPoints, message);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class RentResponse {\n");
    
    sb.append("    filmId: ").append(toIndentedString(filmId)).append("\n");
    sb.append("    pay: ").append(toIndentedString(pay)).append("\n");
    sb.append("    bonusPoints: ").append(toIndentedString(bonusPoints)).append("\n");
    sb.append("    message: ").append(toIndentedString(message)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}